compila mas nao funcina

+++ src/printxnbr.c
@@ -12,19 +12,19 @@
 
 #include "../include/ft_printf.h"
 
-int printxnbr(unsigned int nb)
+int	printxnbr(unsigned long nb)
 {
-    char    *base;
-    int     count;
+	char	*base;
+	int		count;
 
-    count = 0;
-    base = "0123456789abcdef";
-    while (nb >= 16)
-    {
-        count += write(1,&base[nb / 16], 1);
-        nb = nb % 16;
-    }
-    if (nb < 16)
-        count += write(1, &base[nb], 1);
-    return (count); 
+	count = 0;
+	base = "0123456789abcdef";
+	while (nb >= 16)
+	{
+		count += write(1, &base[nb / 16], 1);
+		nb = nb % 16;
+	}
+	if (nb < 16)
+		count += write(1, &base[nb], 1);
+	return (count);
 }
